requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
gspread>=5.12.0
python-dotenv>=1.0.0
email-validator>=2.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.5.0
